[EMCAL-524, EMCAL-526] New histograms + fixes for input handling in downscaled modes (#800)
* [EMCAL-524] added 4 Digits+Raw histos(JIRA 04/Aug/21)

* [EMCAL-526] Fix filter for empty timeframes in case of downsampling

- Channel type changing when applying
  downsampling, filter needs to be done based on the
  binding
- In addition some range fixes

* [EMCAL-524] Fix handling of event building in case of downsampling

Cell and trigger record container cannot be filtered
according to the channel type as it changes in case
the a data sampling policy is applied. Also for
wildcarded inputs the iteration must be done but
based on the binding. As no matcher exist for the
binding the iteration is done by hand.

Co-authored-by: Cristina Terrevoli <cristina.terrevoli@cern.ch>