[EMCAL-524] Remove internal digits publisher (#917)
Internal digits publisher was a temporary solution
in order to overcome problems with the workflow
chaining existing in the past. Digits can be
fed into the QC now via the normal publishers
being part of the EMCAL reconstruction workflow.
The internal publisher is therefore no longer
needed but instead creates trouble in updating
the publisher due to synchronization of the two
repositories.