If possible, use --id argument instead runQC as IL facility (#1681)
This a rather hacky way to retrieve the identity of a Data Processor before it defines its workflow.

This does not work, probably because "--id" is not one of the expected parameters for the user workflow:
```
auto id = config.options().hasOption("id") && config.options().isSet("id") ? config.options().get<std::string>("id") : "runQC";
```

While this improves the situation, we will still have e.g. "qc-task-ITS-ITSFEE" (device id) at the beginning of the task lifetime, but "task/ITSFEE" once it fully initializes. We could consider using device ID throughout the lifetime, to be consistent.