Do not check empty MOs, sanitize testCheck, compute aggregated Activity for QO (#1701)
I added a protection to avoid checking MOs which are nullptr, which testCheck previously abused for no good reason.

Also, most of the tests in testCheck did not have any sense, they tested a dummy class that was written in testCheck. I replaced them with something more meaningful.

The main point of this commit is that now Check computes the aggregated activity, though it is not used yet to store objects, this will come later.