[QC-1023] Legend customizations in SlicedTrendingTask (#1974)
* [QC-1023] Add simple legend customization

Customizable settings:
- Number of columns
- Text size

* [QC-1023] Fix range string Y axis

Since the loop over the y slices is nested within
the loop over the x slices, the way the title was
constructed was leading to appending the new
y-range to previous y-ranges. Therefore later
slices contained all previous slices in y. The
x-range must be cached, and in the iteration of
the y-range for each slice a new title must be
constructed starting from the cached x-title.

* [QC-1023] Configurable legend entries

Default legend titles are replaced by custom legend
titles which are constructed from the ranges in the
raw titles. Customization (both in x and y)
- observable name
- observable unit (optional)
- centmode (mostly for indexed observables)