QC-1142 fix object preservation for 1_per_run policy (#2228)
In some cases (e.g. a postprocessing task trending moving windows), the "validFrom" of the created objects might evolve slightly to the past.
This is what we may see at the beginning of run:

I	11	09:25:34.936	check/sink-PTOF_TOFTrendingO1e7a	TOF	2mZYw7qfc8t	549286	Validity of MO 'TOFTrendingOrbits' is (1711956219957, 1711956273223)
I	11	09:26:34.055	check/sink-PTOF_TOFTrendingO1e7a	TOF	2mZYw7qfc8t	549286	Validity of MO 'TOFTrendingOrbits' is (1711956219854, 1711956333793)
I	11	09:27:34.090	check/sink-PTOF_TOFTrendingO1e7a	TOF	2mZYw7qfc8t	549286	Validity of MO 'TOFTrendingOrbits' is (1711956219854, 1711956392709)

On the other hand, the 1_per_run policy tries to sort the objects by "validFrom" and only if they are equal, it sorts them by "createdAt".
Thus, in the discussed case it will preserve the object which is the latest in terms of "validForm", thus the object from the beginning of the run.

This commit fixes 1_per_run policy by relying solely on "createdAt" to find the latest object for a run.
I could not think of any cases when this could not be applied.