[Common] allow ignoring period and pass when retrieving reference plots (#2381)
The period and pass names were always included in the QCDB queries for
retrieving the reference plots. This means that a run from a given
period and pass could not be used as reference for a different period
and/or pass, which is a strong limitation that needs to be made configurable.

The new code adds two configuration parameters that control whether the
period and/or pass names should be ignored or not in the queries (see
documentation for details).