Fix CompletionPolicy for BookkeepingQualitySink (#2465)
A label for matching the correct CompletionPolicy in BookkeepingQualitySink was missing. This was causing it to wait indefinitely for matching inputs for the same timeslice if QOs from different Data Processors were requested.
The change allows BookkeepingQualitySink to consume the inputs immediatelly and pass them to any other process which may also need them.