Fix TrendCheck creating Quality without a name (#2584)
A default Quality would be created when TrendCheck::beautify was trying to access a non-existent entry in mQualities map.
When using operator[] with a missing key, the object is added using the default constructor, so Quality(10, "") was being created.
This was causing crashes in QualityTask, which was relying on a non-empty Quality name to be there.